Ivan (#16)
* stablishing basic menu

* improving menus

* menu changes

* final menu ended, allow comparation between direct and indirect flights

* last conflict resolution and reformat.

* definitive commit xd

+++ FlightFinder/src/main/java/com/solvd/flightfinder/entities/Country.java
@@ -3,37 +3,37 @@ package com.solvd.flightfinder.entities;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class Country {
-	@JsonProperty
-	private long countryId;
-	@JsonProperty
-	private String name;
+    @JsonProperty
+    private long countryId;
+    @JsonProperty
+    private String name;
 
-	public Country() {
-	}
+    public Country() {
+    }
 
-	public Country(String name) {
-		this.name = name;
-	}
+    public Country(String name) {
+        this.name = name;
+    }
 
-	public float getCountryId() {
-		return countryId;
-	}
+    public long getCountryId() {
+        return countryId;
+    }
 
-	public void setCountryId(long countryId) {
-		this.countryId = countryId;
-	}
+    public void setCountryId(long countryId) {
+        this.countryId = countryId;
+    }
 
-	public String getName() {
-		return name;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public void setName(String name) {
-		this.name = name;
-	}
+    public void setName(String name) {
+        this.name = name;
+    }
 
-	@Override
-	public String toString() {
-		return "Country [countryId=" + countryId + ", name=" + name + "]";
-	}
+    @Override
+    public String toString() {
+        return "Country [countryId=" + countryId + ", name=" + name + "]";
+    }
 
 }
